+++ lib/components/desktop/desktop.dart
@@ -1,0 +1,4 @@
+export "desktop_home.dart";
+export "desktop_projects.dart";
+export "desktop_contents.dart";
+export "desktop_contact.dart";