+++ lib/widgets/widgets.dart
@@ -1,6 +1,5 @@
 export "animations/animations.dart";
 export "background/background.dart";
-export "body/body.dart";
 export "images/images.dart";
 export "interactables/interactables.dart";
 export "layout/layout.dart";